requirements.txt
accelerate==1.10.0
autoawq==0.2.6 
autoawq-kernels==0.0.7
dashscope==1.24.2
openai==1.105.0
transformers==4.45.2
pandas==2.3.2
tiktoken==0.7.0
jinja2==3.1.6
matplotlib==3.10.6
mcp==1.15.0
modelscope==1.18.1
ms-swift==2.4.2
nltk==3.9.1
scikit-learn==1.7.1
ipywidgets==8.1.7
llama-index-embeddings-dashscope==0.1.4
llama-index-llms-openai-like==0.1.3
llama-index-llms-langchain==0.3.0
llama-index-readers-file==0.1.33
seaborn==0.13.2
langchain_openai==0.1.25
langchain_dashscope==0.1.8
langchain_community==0.2.17
ragas==0.1.9
PyMuPDF==1.26.4
pymupdf4llm==0.0.27
tabulate==0.9.0
tzlocal==5.3.1